+++ ChessViewer.Domain/GameMoveModel.cs
@@ -4,9 +4,9 @@ namespace ChessViewer.Domain
 {
     public class GameMoveModel
     {
-        public int Id { get; set; }
-        public int GameId { get; set; }
-        public int MoveNumber { get; set; }
+        public long Id { get; set; }
+        public long GameId { get; set; }
+        public long MoveNumber { get; set; }
         public string WhiteFrom { get; set; }
         public string WhiteTo { get; set; }
         public string BlackFrom { get; set; }
